Fix tox test loop in mvn-phase-script.sh
This loop was not running correctly, because the find results were
being returned as a string rather than an array. The results always
contain a newline, and the complicated rev-cut-rev script seems to
be a workaround that fails if the wrong result is shown first.

By adding parentheses around the TOXINIS assignment command, the
results are returned as an array. We can then simply use bash's
dirname command to get the parent directory.

Issue-ID: MODELING-604
Signed-off-by: Eric Ball <eball@linuxfoundation.org>
Change-Id: I0881958395c7104f263928a3c4394fd141b79208